requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.13
async-timeout==4.0.2
attrs==22.1.0
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.6
dj-database-url==1.0.0
Django==4.1.4
django-filter==22.1
django-on-heroku==1.1.2
djangorestframework==3.14.0
filelock==3.8.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isort==5.10.1
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
kombu==5.2.4
lazy-object-proxy==1.8.0
lxml==4.9.1
Markdown==3.4.1
mccabe==0.7.0
packaging==21.3
parsel==1.7.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
Protego==0.2.1
psycopg2-binary==2.9.5
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pylint==2.15.8
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
queuelib==1.6.2
redis==4.4.0
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
scrapy-random-useragent==0.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tomli==2.0.1
tomlkit==0.11.6
Twisted==22.10.0
typing_extensions==4.4.0
upgrade-requirements==1.7.0
urllib3==1.26.13
vine==5.0.0
w3lib==2.1.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zope.interface==5.5.2
